+++ api/src/main/java/net/jqwik/api/arbitraries/CharacterArbitrary.java
@@ -20,12 +20,20 @@ public interface CharacterArbitrary extends Arbitrary<Character> {
 	CharacterArbitrary all();
 
 	/**
-	 * Allow all chars in {@code chars} show up in generated values.
+	 * Allow all chars in {@code allowedChars} show up in generated values.
 	 *
 	 * Adds to all already allowed chars.
 	 */
 	@API(status = MAINTAINED, since = "1.1.3")
 	CharacterArbitrary with(char... allowedChars);
+
+	/**
+	 * Allow all chars in {@code allowedChars} show up in generated values.
+	 *
+	 * Adds to all already allowed chars.
+	 */
+	@API(status = MAINTAINED, since = "1.2.1")
+	CharacterArbitrary with(CharSequence allowedChars);
 
 	/**
 	 * Allow all chars within {@code min} (included) and {@code max} (included) to show up in generated values.